video: pool: Fix pool size configuration for DMA DRM
Now that VideoPool accept DMA DRM, it is quite likely that the size does not get
provided in the configuration even if the format have a matching GstVideoFormat.

Fix this by reporting back the minimum expected size in the configuration. This
allows the configuration handshake to succeed.

Fixes regression caused by !9345

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9600>