glimagesink: Clip texture if its bigger then display
When handling external textures, there is cases where the texture needs to be
imported in its padded dimensions. The side effect is a larger texture that we
must clip to avoid showing random data found in the padding zone.

This area is clipped by modifying the U/V component of the vertices buffer.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/10120>